feat: adding session blocker
